requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
netmiko==4.1.2
python-dotenv==1.0.0
sqlalchemy==2.0.12
pydantic==1.10.7
jinja2==3.1.2
